chore: move remaining webdriver commands to own interface
